Reworked UI of Fuseki, but *only* for windows.
(Hopefully) made it such that fuseki will never try and use a UNIX socket, nor will it try and use a tcp socket for soliciting, or a udp socket for joining.

Additionally, the interface is (I think) much more streamlined, and the help is more helpful. 

You may be thinking "if this streamlined interface is so much better, why didn't he put it in the UNIX compile?", to which I reply that *I* think it is better, and I am accustomed to windows interfaces, so a linux user might find it ghastly. 

It will almost certainly not work properly, due to my lack of understanding of John's options parser. One thing I am slightly worried about, it seems quite capable of binding to such things as "udp|::GHfiosfhk", but that is probably a glitch in the network code, if it is a glitch at all. 

Simeon

git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@314 6f43db7c-3df8-0310-95f7-eae406835aab