More progress on building Perl bindings. The new procedure is as follows: cd bindings/perl5/libsakusen make ./Makefile.PL
This overwrites the Makefile from the repository. You can then go on to invoke make again, which will actually build the module. make test runs the tests under the harness, and make dist tars up the module for distribution.

Problems:
We really don't want to overwrite the Makefile from the repo every time. We have the choice of moving the code to call swig (which is all we use out of that Makefile) into a Makefile with a different name, or move it out of the directory completely into bindings/perl5. Both approaches have things going for them so I haven't decided on one yet. It seems to be quite hard to get Makefile.PL to run swig, but I haven't given up yet.

We still aren't using libtool. Being able to install and package the module is of less use when it won't work anywhere else :->

It needs plumbed in s.t. you can configure WITH_PERL or such in the top-level configuration and have it automatically build the Perl bindings, and optionally test them. This should be standard across all bindings.


git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@367 6f43db7c-3df8-0310-95f7-eae406835aab