Several changes to Point. Notably, get rid of the specialisation of Point<double>, which was causing lots of code duplication, and the various versions were getting out of sync. truncate() and round() are now declared for all specialisations of Point, but only defined for Point<double>, so using them on any other Point will get you a compile-time error. They no longer cast to a given integer type, so you have to use the normal type conversion mechanism as well, just as when trunc()'ing floats.
Also, the mess with various incompatible operator* and / has been fixed. There
is now a single definition of each that takes a Point<T> and returns a
Point<T>, and all conversions are done with the usual type conversion
mechanism. Thus, if you want the fairly common case of scaling a Point<some
integer> p with a double d, you should do (Point<double>(p) * d).

I was considering a template<U> Point<T>::operator*(U), but in such a case it
is not clear whether the Point<T> should be converted to a Point<U> (as in the
example above), or the U be converted to a T (as when multiplying by a narrower
integer type, or a Point<double> by an integer), so I make you specify.
Possibly, it might be achieved by explicit specialisations, but let's not go
into that yet.

In addition, I've added typemunger.h which should help to sort out some
template woes in the long term. It's not yet tested or documented so will
probably need tweaking.

Finally, I've updated the doxymentation for Visibility to sync with the change
from uint8's to sint8's.


git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@381 6f43db7c-3df8-0310-95f7-eae406835aab