Build fixes for MSVC. Nothing very major.
This time I made libsakusen a static library instead of a dll. This makes my life easier, since I no longer have to mess around with dllexport errors, explicit template instantiation, etc, etc. I have left the stuff that was there to make it partially work in, so it can be reverted if we find a reason to do so.

One bunch of warnings remain, awaiting the result of z1's cogitations for the commit.

git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@490 6f43db7c-3df8-0310-95f7-eae406835aab