Get rid of the compiler-specific guards in maskedptr.h.
It turns out that they are unnecessary: hash<T> is a struct in the g++ STL
as well as in STLport (as they are both essentially the SGI codebase), but
g++ doesn't care if you define as a class something that was declared a
struct, unlike MSVC.

This commit also replaces tabs with spaces. Simeon, please make your editor
not write tabs.

Signed-off-by: Daniel Hulme <st@istic.org>


git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@493 6f43db7c-3df8-0310-95f7-eae406835aab