Replace typedef'd HitPoints with actual type.
As part of my program of not introducing semantics without introducing types,
I've made HitPoints a fully fledged type. This isn't just a type to aid
type-safety: it also introduces saturating arithmetic, and removes the special
meaning of -1u. The major programmer-visible change is that you can't just pass
an integer as a HitPoints now, you have to explicitly call the ctor.

Signed-off-by: Daniel Hulme <st@istic.org>