Revert Simeon's addition of conflict markers in tedomari. Introduced in 3d7841d4 (r537).
Note that the #ifdef'ing out of cairoui.cpp breaks the build if DISABLE_CAIRO
is set, as it produces empty translation units, which are illegal in C and C++.
Such files should instead be added to EXCLUDE_FILES in the Makefile. I leave it
to Simeon to decide what to do in this case.

Signed-off-by: Daniel Hulme <st@istic.org>