Add `make install` action. Doesn't entirely work yet.
It will install binaries and libraries, but no manpages or other data. Plugins
should by rights be in $PREFIX/lib/sakusen, but are in $PREFIX/lib because you
can't override actions the way round you need to. I remind people who are
hoping for packages that a working install procedure is a prerequisite for
same. I'd like to also draw users' attention to the difference between PREFIX
and DESTDIR.

Signed-off-by: Daniel Hulme <st@istic.org>