Disconnect signal.
This may seem like a step backwards (and indeed it is),but I think
listener doesn't really deserve to be a class at all. How I'm going
to give user code access to updates I'm not quite sure, but it may
well end up not being through the Qt signal/slot mechanism. Or possibly
I'll add those back in when the design is cleaner.
Signed-off-by: Michael Donaghy <md401@tommy.(none)>