Use system libltdl by default
The hacked version of libltdl bundled with sakusen was supposed to be a
temporary workaround until mainline ltdl fixed its error messages.  That
proved prohibitively difficult so it stuck.  Nevertheless, it's not very
portable, so I've changed the code to use the system ltdl by default, and only
use the hacked version on explicit request (see INSTALL for more info).

Signed-off-by: John Bytheway <jbytheway@gmail.com>