Rewrite bottomNumber
The previous implementation was generating a compiler warning due to
negating an unsigned number.  I think this is the same warning that MSVC
used to give for which there was already a workaround, but I decided to
avoid it by rewriting the function so that the code which already should
never have run is now never compiled.

This change would be much neater with access to C++17's if constexpr
feature.