Remove BOOST_LIB_SUFFIX configuration option
This was added presumably for a system where the -mt versions of the
Boost libraries were the only ones installed.  But the feature had
rotted, indicating that it's probably not needed any more.

We will need to worry about this again if we want multi-threaded
execution but for now just get rid of it.

Fixes #12